services/prusaconnect/requirements.txt
requests
influxdb
python-dotenv
